requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
python-dotenv==1.0.1

SQLAlchemy==2.0.31
PyMySQL==1.1.1
alembic==1.13.2

pandas==2.2.2
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
